backend/requirements.txt
django>=5,<6
djangorestframework>=3.17
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
psycopg[binary]>=3.3
python-dotenv>=1.2

